src/pyproject.toml
[tool.poetry]
name = "gitlab-due-sync-src"
version = "0.1.0"
description = ""
authors = ["Sarisia <sarisiaice@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.27.1"
boto3 = "^1.23.10"
google-api-python-client = "^2.49.0"
google-auth-oauthlib = "^0.5.1"

[tool.poetry.dev-dependencies]
boto3-stubs = {extras = ["dynamodb"], version = "^1.23.10"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
